fix: change create pair button place (#200)
