+++ src/components/Pools/Sidebar/index.tsx
@@ -1,7 +1,8 @@
-import React, { useCallback, useState } from 'react';
-// import { Text, Box, useTranslation, PoolImportModal } from '@pangolindex/components'
+import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Box, Text } from 'src/components';
+import useParsedQueryString from 'src/hooks/useParsedQueryString';
+import AddLiquidityModal from '../AddLiquidityModal';
 import PoolImportModal from '../PoolImportModal';
 import { Circle, Menu, MenuItem, MenuLink, MenuName, SidebarWrapper } from './styleds';
 
@@ -22,6 +23,9 @@ interface MenuProps {
 const Sidebar = ({ setMenu, activeMenu, menuItems, onManagePoolsClick }: MenuProps) => {
   const { t } = useTranslation();
   const [isPoolImportModalOpen, setIsPoolImportModalOpen] = useState(false);
+  const [isAddLiquidityModalOpen, setAddLiquidityModalOpen] = useState<boolean>(false);
+
+  const parsedQs = useParsedQueryString();
 
   const handlePoolImportModalClose = useCallback(() => {
     setIsPoolImportModalOpen(false);
@@ -36,6 +40,18 @@ const Sidebar = ({ setMenu, activeMenu, menuItems, onManagePoolsClick }: MenuPro
     onManagePoolsClick();
   }, [setIsPoolImportModalOpen, onManagePoolsClick]);
 
+  const handleAddLiquidityModalClose = useCallback(() => {
+    setAddLiquidityModalOpen(false);
+  }, [setAddLiquidityModalOpen]);
+
+  const currency0 = parsedQs?.currency0;
+  const currency1 = parsedQs?.currency1;
+
+  useEffect(() => {
+    if (currency0 && currency1) {
+      setAddLiquidityModalOpen(true);
+    }
+  }, [currency0, currency1]);
   return (
     <SidebarWrapper>
       <Text color="text1" fontSize={[32, 28]} fontWeight={500} ml={20} mt={10}>
@@ -56,7 +72,13 @@ const Sidebar = ({ setMenu, activeMenu, menuItems, onManagePoolsClick }: MenuPro
       </Menu>
 
       <Box padding="8px" mb={10} ml="12px">
-        <Text color="color6" fontSize={14}>
+        <Text fontSize={14} color="primary" onClick={() => setAddLiquidityModalOpen(true)} cursor="pointer">
+          {t('navigationTabs.createPair')}
+        </Text>
+      </Box>
+
+      <Box padding="8px" mb={10} ml="12px">
+        <Text color="color6" fontSize={14} mb="5px">
           {t('pool.noSeePoolJoined')}
         </Text>
 
@@ -64,7 +86,7 @@ const Sidebar = ({ setMenu, activeMenu, menuItems, onManagePoolsClick }: MenuPro
           {t('pool.importIt')}
         </Text>
       </Box>
-
+      <AddLiquidityModal isOpen={isAddLiquidityModalOpen} onClose={handleAddLiquidityModalClose} />
       <PoolImportModal
         isOpen={isPoolImportModalOpen}
         onClose={handlePoolImportModalClose}
