+++ src/components/Pools/index.tsx
@@ -3,7 +3,6 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { BIG_INT_ZERO } from 'src/constants';
 import { useChainId } from 'src/hooks';
-import useParsedQueryString from 'src/hooks/useParsedQueryString';
 import { usePangoChefInfosHook } from 'src/state/ppangoChef/multiChainsHooks';
 import { PangoChefInfo } from 'src/state/ppangoChef/types';
 import {
@@ -12,33 +11,20 @@ import {
   useMinichefStakingInfosHook,
 } from 'src/state/pstake/multiChainsHooks';
 import { MinichefStakingInfo, PoolType } from 'src/state/pstake/types';
-import { Hidden } from 'src/theme/components';
 import { isEvmChain } from 'src/utils';
 import { Box } from '../Box';
-import AddLiquidityModal from './AddLiquidityModal';
 import Pool from './Pool';
 import Sidebar, { MenuType } from './Sidebar';
 import Wallet from './Wallet';
-import { ExternalLink, GridContainer, PageWrapper } from './styleds';
+import { GridContainer, PageWrapper } from './styleds';
 
 const PoolsUI = () => {
   const chainId = useChainId();
   const minichef = CHAINS[chainId].contracts?.mini_chef;
 
   const [activeMenu, setMenu] = useState<string>(MenuType.yourPool);
-  const [isAddLiquidityModalOpen, setAddLiquidityModalOpen] = useState<boolean>(false);
+
   const { t } = useTranslation();
-
-  const parsedQs = useParsedQueryString();
-
-  const currency0 = parsedQs?.currency0;
-  const currency1 = parsedQs?.currency1;
-
-  useEffect(() => {
-    if (currency0 && currency1) {
-      setAddLiquidityModalOpen(true);
-    }
-  }, [currency0, currency1]);
 
   const useGetAllFarmData = useGetAllFarmDataHook[chainId];
 
@@ -46,10 +32,6 @@ const PoolsUI = () => {
   const pangoChefStakingInfos = usePangoChefInfosHook[chainId]() || [];
   const subgraphMiniChefStakingInfo = useGetMinichefStakingInfosViaSubgraphHook[chainId]() || [];
   const onChainMiniChefStakingInfo = useMinichefStakingInfosHook[chainId]() || [];
-
-  const handleAddLiquidityModalClose = useCallback(() => {
-    setAddLiquidityModalOpen(false);
-  }, [setAddLiquidityModalOpen]);
 
   // filter only live or needs migration pools
   const miniChefStakingInfo = useMemo(() => {
@@ -206,15 +188,7 @@ const PoolsUI = () => {
             <Wallet activeMenu={activeMenu} setMenu={handleSetMenu} menuItems={menuItems} />
           )}
         </Box>
-        <Hidden upToSmall={true}>
-          <Box>
-            <ExternalLink onClick={() => setAddLiquidityModalOpen(true)} style={{ cursor: 'pointer' }}>
-              {t('navigationTabs.createPair')}
-            </ExternalLink>
-          </Box>
-        </Hidden>
       </GridContainer>
-      <AddLiquidityModal isOpen={isAddLiquidityModalOpen} onClose={handleAddLiquidityModalClose} />
     </PageWrapper>
   );
 };
