+++ src/components/Pools/styleds.tsx
@@ -9,7 +9,7 @@ export const PageWrapper = styled(Box)`
 
 export const GridContainer = styled(Box)`
   display: grid;
-  grid-template-columns: minmax(auto, 75%) minmax(auto, 25%);
+  grid-template-columns: 100%;
   grid-gap: 12px;
   padding: 50px 0px 0px;
   height: 100%;
@@ -17,16 +17,4 @@ export const GridContainer = styled(Box)`
     padding: 10px 0px 0px;
     grid-template-columns: 100%;
   `};
-`;
-
-export const ExternalLink = styled.a`
-  background-color: ${({ theme }) => theme.bg2};
-  color: ${({ theme }) => theme.text1};
-  display: flex;
-  text-decoration: none;
-  height: 80px;
-  align-items: center;
-  justify-content: center;
-  border-radius: 10px;
-  margin-bottom: 15px;
 `;
